# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prueba Resuelve",
        "description": "Prueba Resuelve",
        "contact": {
            "email": "luisvl13@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/niveles": {
            "get": {
                "tags": [
                    "Niveles"
                ],
                "summary": "Obtener lista de Niveles",
                "operationId": "1518053541d6fdd54013ba811f65c5cb",
                "responses": {
                    "200": {
                        "description": "Lista de los niveles y sus niveles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "messages": {
                                            "type": "string",
                                            "example": "Operación realizada con exito"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nivel": {
                                                        "type": "string",
                                                        "example": "A"
                                                    },
                                                    "goles_mes": {
                                                        "type": "number",
                                                        "example": "5"
                                                    },
                                                    "equipos_id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "404"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Niveles"
                ],
                "summary": "Crea un nivel",
                "operationId": "45a33ec6721fe598589c82d81b865a99",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nivel": {
                                                "type": "string"
                                            },
                                            "goles_mes": {
                                                "type": "integer"
                                            },
                                            "equipos_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nivel": "A",
                                    "goles_mes": 5,
                                    "equipos_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nivel creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "messages": {
                                            "type": "string",
                                            "example": "Registro creado"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nivel": {
                                                        "type": "string",
                                                        "example": "A"
                                                    },
                                                    "goles_meta": {
                                                        "type": "integer",
                                                        "example": "5"
                                                    },
                                                    "nombre": {
                                                        "type": "integer",
                                                        "example": "2"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "400"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Conflicto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "500"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}